fix(Academy): SCORING critique — crédit partiel arrondi (1pt 50%=100% bug) + F4 (dédup, tirets, a11y)
Fix systémique : accumulation float du crédit partiel (4 types), percent exact,
passing_score juste. + dédup pool ddwtos, 0 tiret cadratin, MAX_DDWTOS_TEXT/aria.
Attrapé par /sim F4. 669 tests verts (+12).

+++ Modules/Academy/app/Http/Controllers/QuizController.php
@@ -371,7 +371,9 @@ class QuizController extends Controller
         $validated[$index] = [
             'answer'          => $given,
             'correct'         => (bool) ($detail['correct'] ?? false),
-            'points_earned'   => (int) ($single['points_earned'] ?? 0),
+            // Crédit partiel : conserve la valeur native de score() (int si entier,
+            // float ex. 0,5 si fractionnaire) — surtout PAS de cast int qui tronquerait.
+            'points_earned'   => $single['points_earned'] ?? 0,
             'points_possible' => (int) ($single['points_possible'] ?? 0),
         ];
